% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.to.annual.R
\name{ann.value}
\alias{ann.value}
\title{Calculate annual payment from present value

\code{present.to.annual} Compute uniform series annual payments from present value
using capital recovery factor (crf)}
\usage{
ann.value(i, n, PV)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{PV}{present value}
}
\value{
ann.value
}
\description{
Calculate annual payment from present value

\code{present.to.annual} Compute uniform series annual payments from present value
using capital recovery factor (crf)
}
\examples{
# Result: ann.value 
ann.value(0.08, 2000, 30)  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012
}
