% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.to.future.R
\name{annual.to.future}
\alias{annual.to.future}
\title{Calculate future value from uniform annual payment

\code{annual.to.future} Compute future values from uniform annual payments
using uniform series compound amount factor (uscaf)}
\usage{
annual.to.future(i, n, A)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{A}{series of uniform annual payments}
}
\value{
FV
}
\description{
Calculate future value from uniform annual payment

\code{annual.to.future} Compute future values from uniform annual payments
using uniform series compound amount factor (uscaf)
}
\examples{
 
# Result: FV 
annual.to.future(0.08, 30, 50000)  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012
}
