% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectiveI.R
\name{effective_i}
\alias{effective_i}
\title{Calculate effective annual interest rate for a known nominal rate and
compounding period per year

\code{effective_i} Compute future value nominal rate and
compounding period per year, and frequency.}
\usage{
effective_i(r, m = c("yearly", "quarterly", "monthly"))
}
\arguments{
\item{r}{nominal interest rate in decimal number}

\item{m}{number of compounding period per year
monthly = 12
Quarterly = 4
yearly = 1}

\item{i}{interest rate per period (r/m)}
}
\value{
effective_i
}
\description{
Calculate effective annual interest rate for a known nominal rate and
compounding period per year

\code{effective_i} Compute future value nominal rate and
compounding period per year, and frequency.
}
\examples{
 
# Result: Effective_i = 
effective_i (0.08, m = "yearly")  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012.
}
