% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectiveI.R
\name{EffectiveI}
\alias{EffectiveI}
\title{Calculate effective annual interest for a known nominal rate and
compounding period per year

\code{effectiveI} Compute future value nominal rate and
compounding period per year, and frequency.}
\usage{
EffectiveI(r, m = c("yearly", "quarterly", "monthly"))
}
\arguments{
\item{r}{nominal interest rate in decimal number}

\item{m}{number of compounding period per year
monthly = 12
Quarterly = 4
yearly = 1}
}
\value{
effectiveI
}
\description{
Calculate effective annual interest for a known nominal rate and
compounding period per year

\code{effectiveI} Compute future value nominal rate and
compounding period per year, and frequency.
}
\examples{
 
# Result: EffectivI
EffectiveI (0.08, m = "yearly")  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012.
}
