% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.to.annual.R
\name{future.to.annual}
\alias{future.to.annual}
\title{Compute annual payment value from future value

\code{future.to.annual} calculate annual value from future value using
the accumulated amount after years using sinking fund factor (sff)}
\usage{
future.to.annual(i, n, FV)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{FV}{accumulated (future) value}
}
\value{
A
}
\description{
Compute annual payment value from future value

\code{future.to.annual} calculate annual value from future value using
the accumulated amount after years using sinking fund factor (sff)
}
\examples{
 
# Result: A
future.to.annual(0.08, 30, 650000)  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012
}
