% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.to.annual.R
\name{gradient.to.annual}
\alias{gradient.to.annual}
\title{Calculate annual value from uniform gradient payment

\code{gradient.to.future} Compute annual value from uniform gradient payments
using uniform gradient present worth factor (ugaw)}
\usage{
gradient.to.annual(i, n, G)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{G}{uniform gradient payments}
}
\value{
ann.value
}
\description{
Calculate annual value from uniform gradient payment

\code{gradient.to.future} Compute annual value from uniform gradient payments
using uniform gradient present worth factor (ugaw)
}
\examples{
 
# Result: ann.value = $
gradient.to.annual(0.08, 30, )  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012.
}
