% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.to.present.R
\name{gradient.to.present}
\alias{gradient.to.present}
\title{Calculate present value from uniform gradient payment

\code{gradient.to.present} Compute present value from uniform gradient payments
using uniform gradient present worth factor (ugpw)}
\usage{
gradient.to.present(i, n, G)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{G}{uniform gradient payments}
}
\value{
pr.value
}
\description{
Calculate present value from uniform gradient payment

\code{gradient.to.present} Compute present value from uniform gradient payments
using uniform gradient present worth factor (ugpw)
}
\examples{
 
# Result: pr.value =
gradient.to.present(0.08, 30, )  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012.
}
