% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_to_future.R
\name{gradient_to_future}
\alias{gradient_to_future}
\title{Calculate future value from uniform gradient payment

\code{gradient_to_future} Compute future value from uniform gradient payments
using uniform gradient present worth factor (ugfw)}
\usage{
gradient_to_future(i, n, G)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{G}{uniform gradient payments}
}
\value{
fr.value
}
\description{
Calculate future value from uniform gradient payment

\code{gradient_to_future} Compute future value from uniform gradient payments
using uniform gradient present worth factor (ugfw)
}
\examples{
 
# Result: fr.value = $ 312,312
gradient_to_future(0.08, 30,300)  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012.
}
