% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_to_present.R
\name{gradient_to_present}
\alias{gradient_to_present}
\title{Calculate present value from uniform gradient payment

\code{gradient_t_present} Compute present value from uniform gradient payments
using uniform gradient present worth factor (ugpw)}
\usage{
gradient_to_present(i, n, G)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{G}{uniform gradient payments}
}
\value{
pr.value
}
\description{
Calculate present value from uniform gradient payment

\code{gradient_t_present} Compute present value from uniform gradient payments
using uniform gradient present worth factor (ugpw)
}
\examples{
 
# Result: pr.value =
gradient_to_present(0.08, 30, 300 )  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012.
}
