% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest.during.construction.R
\name{interest.during.construction}
\alias{interest.during.construction}
\title{Compute interest during construction

\code{interest.during.construction} calculate interest during construction
using monthly discount factor}
\usage{
interest.during.construction(i, duration, capital)
}
\arguments{
\item{i}{interest rate in percent per year}

\item{duration}{construction duration in months}

\item{capital}{first/capital cost}
}
\value{
idc
}
\description{
Compute interest during construction

\code{interest.during.construction} calculate interest during construction
using monthly discount factor
}
\examples{
 
# Result: idc
interest.during.construction (0.08,25,700000)

#' @export 
}
