% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_during_construction.R
\name{interest_during_construction}
\alias{interest_during_construction}
\title{Compute interest during construction

\code{interest_during_construction} calculate interest during construction
using monthly discount factor}
\usage{
interest_during_construction(i, duration, capital)
}
\arguments{
\item{i}{interest rate in percent per year}

\item{duration}{construction duration in months}

\item{capital}{first/capital cost}
}
\value{
idc
}
\description{
Compute interest during construction

\code{interest_during_construction} calculate interest during construction
using monthly discount factor
}
\examples{
 
# Result: idc
interest_during_construction (0.02758,25,700000)

#' @export 
}
