% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_to_annual.R
\name{present_to_annual}
\alias{present_to_annual}
\title{Calculate annual payment from present value

\code{present_to_annual} Compute uniform series annual payments from present value
using capital recovery factor (crf)}
\usage{
present_to_annual(i, n, PV)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{PV}{present value}
}
\value{
ann.value
}
\description{
Calculate annual payment from present value

\code{present_to_annual} Compute uniform series annual payments from present value
using capital recovery factor (crf)
}
\examples{
# Result: ann.value = $17765
present_to_annual(0.08, 30, 200000)  

}
\references{
Newnan, Donald G., Ted G. Eschenbach, and Jerome P. Lavelle.
Engineering economic analysis. Vol. 12. Oxford University Press, 2012
}
